rearrange compose as one of the backends
