+++ apps/dockup/test/dockup/backends/compose/htpasswd_test.exs
@@ -1,11 +1,11 @@
-defmodule Dockup.HtpasswdTest do
+defmodule Dockup.Backends.Compose.HtpasswdTest do
   use ExUnit.Case, async: true
 
   test "write() writes htpasswd string into both dockup and logio host files" do
     Application.put_env(:dockup, :htpasswd, "foo:bar")
     Application.fetch_env!(:dockup, :htpasswd_dir) |> File.mkdir_p!()
 
-    Dockup.Htpasswd.write()
+    Dockup.Backends.Compose.Htpasswd.write()
 
     htpasswd_dir = Application.fetch_env!(:dockup, :htpasswd_dir)
     dockup_host = Path.join(htpasswd_dir, "127.0.0.1.xip.io")
